--- a/src/core/Akka.MultiNodeTestRunner.Shared/Sinks/TestCoordinatorEnabledMessageSink.cs
+++ b/src/core/Akka.MultiNodeTestRunner.Shared/Sinks/TestCoordinatorEnabledMessageSink.cs
@@ -28,6 +28,7 @@ namespace Akka.MultiNodeTestRunner.Shared.Sinks
             {
                 if (UseTestCoordinator)
                 {
+                    var sender = Sender;
                     TestCoordinatorActorRef.Ask<TestRunTree>(new TestRunCoordinator.RequestTestRunState())
                         .ContinueWith(task =>
                         {
@@ -35,7 +36,7 @@ namespace Akka.MultiNodeTestRunner.Shared.Sinks
                                 ? 0
                                 : 1);
                         }, TaskContinuationOptions.ExecuteSynchronously)
-                            .PipeTo(Sender, Self);
+                            .PipeTo(sender, Self);
                 }
             });
         }