--- a/src/core/Akka.Remote.Tests/RemotingTerminatorSpecs.cs
+++ b/src/core/Akka.Remote.Tests/RemotingTerminatorSpecs.cs
@@ -8,7 +8,9 @@
 using System;
 using Akka.Actor;
 using Akka.Configuration;
+using Akka.Event;
 using Akka.TestKit;
+using Akka.TestKit.TestActors;
 using Xunit;
 
 namespace Akka.Remote.Tests
@@ -65,6 +67,38 @@ namespace Akka.Remote.Tests
             Within(TimeSpan.FromSeconds(10), () =>
             {
                 var terminationTask = sys2.Terminate();
+                Assert.True(terminationTask.Wait(RemainingOrDefault), "Expected to terminate within 10 seconds, but didn't.");
+            });
+        }
+
+        [Fact]
+        public void RemotingTerminator_should_shutdown_properly_with_remotely_deployed_actor()
+        {
+            var sys2 = ActorSystem.Create("System2", RemoteConfig);
+            var sys2Address = RARP.For(sys2).Provider.DefaultAddress;
+            
+            // open an association via remote deployment
+            var associated =
+                Sys.ActorOf(BlackHoleActor.Props.WithDeploy(Deploy.None.WithScope(new RemoteScope(sys2Address))),
+                    "remote");
+            Watch(associated);
+
+            // verify that the association is open (don't terminate until handshake is finished)
+            associated.Ask<ActorIdentity>(new Identify("foo"), RemainingOrDefault).Result.MessageId.ShouldBe("foo");
+            
+            // terminate the DEPLOYED system
+            Within(TimeSpan.FromSeconds(20), () =>
+            {
+                var terminationTask = sys2.Terminate();
+                Assert.True(terminationTask.Wait(RemainingOrDefault), "Expected to terminate within 10 seconds, but didn't.");
+                
+                ExpectTerminated(associated); // expect that the remote deployed actor is dead
+            });
+
+            // now terminate the DEPLOYER system
+            Within(TimeSpan.FromSeconds(10), () =>
+            {
+                var terminationTask = Sys.Terminate();
                 Assert.True(terminationTask.Wait(RemainingOrDefault), "Expected to terminate within 10 seconds, but didn't.");
             });
         }