--- a/src/core/Akka.Streams.TestKit.Tests/StreamTestDefaultMailbox.cs
+++ b/src/core/Akka.Streams.TestKit.Tests/StreamTestDefaultMailbox.cs
@@ -7,7 +7,10 @@
 
 using System;
 using Akka.Actor;
+using Akka.Configuration;
 using Akka.Dispatch;
+using Akka.Dispatch.MessageQueues;
+using Akka.Util.Internal;
 
 namespace Akka.Streams.TestKit.Tests
 {
@@ -16,26 +19,30 @@ namespace Akka.Streams.TestKit.Tests
     /// This mailbox is only used in tests to verify that stream actors are using
     /// the dispatcher defined in <see cref="ActorMaterializerSettings"/>
     /// </summary>
-    public sealed class StreamTestDefaultMailbox : UnboundedMailbox
+    public sealed class StreamTestDefaultMailbox : MailboxType, IProducesMessageQueue<UnboundedMessageQueue>
     {
-        public override void SetActor(ActorCell actorCell)
+
+        public override IMessageQueue Create(IActorRef owner, ActorSystem system)
         {
-            if (actorCell.Self is ActorRefWithCell)
+            if (owner is ActorRefWithCell)
             {
-                var actorType = actorCell.Props.Type;
-                if (!typeof (ActorBase).IsAssignableFrom(actorType))
+                var actorType = owner.AsInstanceOf<ActorRefWithCell>().Underlying.Props.Type;
+                if (!typeof(ActorBase).IsAssignableFrom(actorType))
                     throw new ArgumentException(
-                        string.Format("Don't use anonymouse actor classes, actor class for {0} was [{1}]", actorCell,
-                            actorType.FullName));
+                        $"Don't use anonymouse actor classes, actor class for {owner} was [{actorType.FullName}]");
                 // StreamTcpManager is allowed to use another dispatcher
                 if (actorType.FullName.StartsWith("Akka.Streams."))
                     throw new ArgumentException(
-                        string.Format(
-                            "{0} with actor class [{1}] must not run on default dispatcher in tests. Did you forget to define `props.withDispatcher` when creating the actor? Or did you forget to configure the `akka.stream.materializer` setting accordingly or force the dispatcher using `ActorMaterializerSettings(sys).withDispatcher(\"akka.test.stream-dispatcher\")` in the test?",
-                            actorCell, actorType.FullName));
+                        $"{owner} with actor class [{actorType.FullName}] must not run on default dispatcher in tests. Did you forget to define `props.withDispatcher` when creating the actor? Or did you forget to configure the `akka.stream.materializer` setting accordingly or force the dispatcher using `ActorMaterializerSettings(sys).withDispatcher(\"akka.test.stream-dispatcher\")` in the test?");
             }
+            
+            return new UnboundedMessageQueue();
+        }
 
-            base.SetActor(actorCell);
+        public StreamTestDefaultMailbox() : this(null, null) { }
+
+        public StreamTestDefaultMailbox(Settings settings, Config config) : base(settings, config)
+        {
         }
     }
 }