--- a/src/core/Akka.Tests/Actor/ActorCellSpec.cs
+++ b/src/core/Akka.Tests/Actor/ActorCellSpec.cs
@@ -63,7 +63,7 @@ namespace Akka.Tests.Actor
             var actor = Sys.ActorOf(Props.Create(() => new DummyAsyncActor(autoResetEvent)));
             actor.Tell("hello");
             //ensure the message was received
-            autoResetEvent.WaitOne();
+            Assert.True(autoResetEvent.WaitOne(TimeSpan.FromSeconds(3)), "Timed out while waiting for autoreset event");
             var refCell = actor as ActorRefWithCell;
             var cell = refCell.Underlying as ActorCell;
             //wait while current message is not null (that is, receive is not yet completed/exited)