--- a/src/core/Akka/Actor/Cell.cs
+++ b/src/core/Akka/Actor/Cell.cs
@@ -70,7 +70,7 @@ namespace Akka.Actor
 
         bool IsTerminated { get; }
 
-        void Post(IActorRef sender, object message);
+        void SendMessage(IActorRef sender, object message);