… src/core/Akka.API.Tests/CoreAPISpec.ApproveCore.approved.txt
… src/core/Akka.TestKit/Internal/InternalTestActorRef.cs
… src/core/Akka.Tests.Performance/Dispatch/MailboxMemoryFootprintSpec.cs
… src/core/Akka.Tests/Actor/FSMTimingSpec.cs
… src/core/Akka.Tests/Actor/RootGuardianActorRef_Tests.cs
… src/core/Akka.Tests/Actor/StashMailboxSpec.cs
… src/core/Akka.Tests/Actor/SupervisorHierarchySpec.cs
… src/core/Akka.Tests/Actor/SystemGuardianTests.cs
… src/core/Akka.Tests/Akka.Tests.csproj
… src/core/Akka.Tests/Dispatch/ActorMailboxSpec.cs
… src/core/Akka.Tests/Dispatch/MailboxesSpec.cs
… src/core/Akka.Tests/Dispatch/SysMsg/SystemMessageListSpec.cs
… src/core/Akka.Tests/Event/LoggerMailboxSpec.cs
… src/core/Akka.Tests/Routing/RoutingSpec.cs
… src/core/Akka.Tests/Serialization/SerializationSpec.cs
… src/core/Akka.Tests/Util/IndexSpec.cs
… src/core/Akka/Actor/ActorCell.Children.cs
… src/core/Akka/Actor/ActorCell.DefaultMessages.cs
… src/core/Akka/Actor/ActorCell.FaultHandling.cs
… src/core/Akka/Actor/ActorCell.cs
… src/core/Akka/Actor/ActorRef.cs
… src/core/Akka/Actor/ActorRefProvider.cs
… src/core/Akka/Actor/DeadLetterMailbox.cs
… src/core/Akka/Actor/Internal/ActorSystemImpl.cs
… src/core/Akka/Actor/LocalActorRef.cs
… src/core/Akka/Actor/RepointableActorRef.cs
… src/core/Akka/Actor/RootGuardianActorRef.cs
… src/core/Akka/Actor/Stash/Internal/AbstractStash.cs
… src/core/Akka/Akka.csproj
… src/core/Akka/Configuration/Pigeon.conf