close #2024 - guaranteed that all children are terminated before parent terminates
added mailboxtype; adjusted dispatcher system to match JVM
modified SystemMessages to support SystemMessageList
implemented IRequiresMessageQueue and IProducesMessageQueue