@Aaronontheweb 2016-06-12
Also resolves https://github.com/akkadotnet/akka.net/issues/2028