@Aaronontheweb 2016-06-12
This thing is ready for review - I'll be taking a look at the perf numbers and making some adjustments there.