@Aaronontheweb 2016-06-12
cc @rogeralsing @Horusiath @JeffCyr per what we were discussing in Gitter, I think we're eventually going to have to use something like a lock to fully solve the race issue with the top-level actors. As the first run of the spec showed, even with the CAS stuff I have in place it may not be enough.