@JeffCyr 2016-06-13
@aaronontheweb You need to use AttachChild to create top level actors in ActorSystem.ActorOf