@rogeralsing 2016-06-13
I've gotten through half of the PR so far.
Looks solid so far. will continue later today