@alexvaluyskiy 2016-06-13
@Aaronontheweb I've found a bug with GetRandomActorName
```C#
        [Fact]
        public void Actor_names_must_starts_from_a()
        {
            Sys.ActorOf(Props.Create<EchoActor>(TestActor)).Path.ToStringWithoutAddress().Should().Be("/user/$a");
        }
```
This spec fails, it returns "/user/$b"