@alexvaluyskiy 2016-06-13
@Aaronontheweb why you have not changed Post to SendMessage? We've already have breaking changes here, and will have them in Routing PR.