@Aaronontheweb 2016-06-13
@JeffCyr 

> You need to use AttachChild to create top level actors in ActorSystem.ActorOf

So I took a look at this when I was going through it:

This is `ActorSystemImpl.ActorOf`
```csharp
public override IActorRef ActorOf(Props props, string name = null)
        {
            return _provider.Guardian.Cell.ActorOf(props, name: name);
        }
```

Cell.ActorOf calls

```csharp
private IActorRef ActorOf(Props props, string name, bool isAsync, bool isSystemService)
        {
            if (name == null)
                name = GetRandomActorName();
            else
                CheckName(name);

            return MakeChild(props, name, isAsync, isSystemService);
        }
```

`ActorCell.AttachChild` does the following:

```csharp
public virtual IActorRef AttachChild(Props props, bool isSystemService, string name = null)
        {
            return MakeChild(props, name ?? GetRandomActorName(), true, isSystemService);
        }
```

Basically they both end up doing the same things. I can consolidate these, but the behavior is the same. It's the same with Scala https://github.com/akka/akka/blob/master/akka-actor/src/main/scala/akka/actor/dungeon/Children.scala#L43

```scala
private[akka] def attachChild(props: Props, systemService: Boolean): ActorRef =
    makeChild(this, props, randomName(), async = true, systemService = systemService)
  private[akka] def attachChild(props: Props, name: String, systemService: Boolean): ActorRef =
    makeChild(this, props, checkName(name), async = true, systemService = systemService)
```