@Aaronontheweb 2016-06-13
Only real difference is the extra `CheckName` call in `ActorOf`