@Aaronontheweb 2016-06-13
@alexvaluyskiy 

> why you have not changed Post to SendMessage? We've already have breaking changes here, and will have them in Routing PR.

Eh, fair enough. I'll change that for the sake of consistency with the JVM.