@Aaronontheweb 2016-06-13
@alexvaluyskiy 

> I've found a bug with GetRandomActorName

Hah, I wonder how that happened? Maybe we're starting a counter at 1 instead of at 0 or something like that...