@Aaronontheweb 2016-06-13
Some performance items: in particular, I'm interested in seeing how the dispatcher and mailbox changes might have affected throughput.

Let's do this top-down:

## Before
# Akka.Remote.Tests.Performance.Transports.HeliosRemoteMessagingThroughputSpec+OneWay
__Measures the throughput of Akka.Remote over a particular transport using one-way messaging__
_6/11/2016 8:45:51 PM_
### System Info
```ini
NBench=NBench, Version=0.3.0.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=2
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=2
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=3, MaximumRunTime=00:00:01
Concurrent=True
Tracing=True
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |          191.00 |          184.00 |          174.00 |            8.89 |
TotalCollections [Gen1] |     collections |           11.00 |           10.00 |            9.00 |            1.00 |
TotalCollections [Gen2] |     collections |            1.00 |            1.00 |            1.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |       50,000.00 |       50,000.00 |       50,000.00 |            0.00 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |           74.76 |           73.08 |           70.94 |            1.95 |
TotalCollections [Gen1] |     collections |            4.33 |            3.98 |            3.52 |            0.41 |
TotalCollections [Gen2] |     collections |            0.41 |            0.40 |            0.39 |            0.01 |
[Counter] RemoteMessageReceived |      operations |       20,386.25 |       19,873.65 |       19,570.59 |          446.38 |


## After 
# Akka.Remote.Tests.Performance.Transports.HeliosRemoteMessagingThroughputSpec+OneWay
__Measures the throughput of Akka.Remote over a particular transport using one-way messaging__
_6/12/2016 10:59:30 PM_
### System Info
```ini
NBench=NBench, Version=0.3.0.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=2
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=2
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=3, MaximumRunTime=00:00:01
Concurrent=True
Tracing=True
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |          190.00 |          179.67 |          169.00 |           10.50 |
TotalCollections [Gen1] |     collections |           15.00 |           12.00 |            8.00 |            3.61 |
TotalCollections [Gen2] |     collections |            3.00 |            2.67 |            2.00 |            0.58 |
[Counter] RemoteMessageReceived |      operations |       50,000.00 |       50,000.00 |       50,000.00 |            0.00 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |          102.19 |           98.06 |           92.02 |            5.35 |
TotalCollections [Gen1] |     collections |            7.89 |            6.54 |            4.36 |            1.91 |
TotalCollections [Gen2] |     collections |            1.70 |            1.46 |            1.09 |            0.32 |
[Counter] RemoteMessageReceived |      operations |       28,386.28 |       27,305.28 |       26,305.83 |        1,042.62 |

The most recent perf runs on the `dev` branch have all placed the Helios transport's throughput during one-way messaging in the 20k-22k msg / s range. This spec has it clocking in between 26.3-28.3k msg / s. Looks promising!