@Aaronontheweb 2016-06-13
Now looking at *just* the `ForkJoinDispatcher` running the normal remoting pipeline:

## Before
# Akka.Remote.Tests.Performance.ForkJoinDispatcherRemoteMessagingThroughputSpec+OneWay
__Measures the throughput of Akka.Remote over a particular transport using one-way messaging__
_6/11/2016 8:44:31 PM_
### System Info
```ini
NBench=NBench, Version=0.3.0.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=2
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=2
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=3, MaximumRunTime=00:00:01
Concurrent=True
Tracing=True
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |          118.00 |          117.33 |          117.00 |            0.58 |
TotalCollections [Gen1] |     collections |            7.00 |            7.00 |            7.00 |            0.00 |
TotalCollections [Gen2] |     collections |            2.00 |            1.33 |            1.00 |            0.58 |
[Counter] RemoteMessageReceived |      operations |       50,000.00 |       50,000.00 |       50,000.00 |            0.00 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |           99.17 |           97.34 |           95.45 |            1.86 |
TotalCollections [Gen1] |     collections |            5.88 |            5.81 |            5.71 |            0.09 |
TotalCollections [Gen2] |     collections |            1.68 |            1.11 |            0.82 |            0.49 |
[Counter] RemoteMessageReceived |      operations |       42,020.92 |       41,477.23 |       40,788.96 |          628.58 |

### Raw Data
#### TotalCollections [Gen0]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |          118.00 |           99.17 |   10,083,759.32 |
               2 |          117.00 |           97.40 |   10,267,462.39 |
               3 |          117.00 |           95.45 |   10,477,110.26 |

#### TotalCollections [Gen1]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            7.00 |            5.88 |  169,983,371.43 |
               2 |            7.00 |            5.83 |  171,613,300.00 |
               3 |            7.00 |            5.71 |  175,117,414.29 |

#### TotalCollections [Gen2]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            2.00 |            1.68 |  594,941,800.00 |
               2 |            1.00 |            0.83 |1,201,293,100.00 |
               3 |            1.00 |            0.82 |1,225,821,900.00 |

#### [Counter] RemoteMessageReceived
           Run # |      operations |  operations / s | ns / operations |
---------------- |---------------- |---------------- |---------------- |
               1 |       50,000.00 |       42,020.92 |       23,797.67 |
               2 |       50,000.00 |       41,621.82 |       24,025.86 |
               3 |       50,000.00 |       40,788.96 |       24,516.44

## After
# Akka.Remote.Tests.Performance.ForkJoinDispatcherRemoteMessagingThroughputSpec+OneWay
__Measures the throughput of Akka.Remote over a particular transport using one-way messaging__
_6/12/2016 10:58:22 PM_
### System Info
```ini
NBench=NBench, Version=0.3.0.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=2
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=2
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=3, MaximumRunTime=00:00:01
Concurrent=True
Tracing=True
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |          118.00 |          117.33 |          117.00 |            0.58 |
TotalCollections [Gen1] |     collections |            7.00 |            7.00 |            7.00 |            0.00 |
TotalCollections [Gen2] |     collections |            2.00 |            1.33 |            1.00 |            0.58 |
[Counter] RemoteMessageReceived |      operations |       50,000.00 |       50,000.00 |       50,000.00 |            0.00 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |          113.05 |          110.44 |          107.68 |            2.69 |
TotalCollections [Gen1] |     collections |            6.71 |            6.59 |            6.44 |            0.13 |
TotalCollections [Gen2] |     collections |            1.92 |            1.26 |            0.92 |            0.57 |
[Counter] RemoteMessageReceived |      operations |       47,903.78 |       47,061.55 |       46,016.56 |          959.81 |

### Raw Data
#### TotalCollections [Gen0]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |          117.00 |          110.60 |    9,041,712.82 |
               2 |          117.00 |          107.68 |    9,286,882.91 |
               3 |          118.00 |          113.05 |    8,845,415.25 |

#### TotalCollections [Gen1]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            7.00 |            6.62 |  151,125,771.43 |
               2 |            7.00 |            6.44 |  155,223,614.29 |
               3 |            7.00 |            6.71 |  149,108,428.57 |

#### TotalCollections [Gen2]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            1.00 |            0.95 |1,057,880,400.00 |
               2 |            1.00 |            0.92 |1,086,565,300.00 |
               3 |            2.00 |            1.92 |  521,879,500.00 |

#### [Counter] RemoteMessageReceived
           Run # |      operations |  operations / s | ns / operations |
---------------- |---------------- |---------------- |---------------- |
               1 |       50,000.00 |       47,264.32 |       21,157.61 |
               2 |       50,000.00 |       46,016.56 |       21,731.31 |
               3 |       50,000.00 |       47,903.78 |       20,875.18 |

Previously clocked in between 40.8k - 42k msg / s. Doing about 46-48k msg / s now. Seems consistent with the performance change we saw when using Helios.