@Aaronontheweb 2016-06-13
Now let's look at the `ReceiveActor` throughput numbers:

## Before
# Akka.Tests.Performance.Actor.ReceiveActorThroughputSpec+Benchmark
__Measures the throughput of an actor implementation__
_6/11/2016 8:57:44 PM_
### System Info
```ini
NBench=NBench, Version=0.3.0.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=2
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=2
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=13, MaximumRunTime=00:00:01
Concurrent=True
Tracing=True
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |           55.00 |           53.46 |           52.00 |            1.05 |
TotalCollections [Gen1] |     collections |           24.00 |           20.46 |           18.00 |            1.56 |
TotalCollections [Gen2] |     collections |            7.00 |            5.38 |            4.00 |            0.96 |
[Counter] MessageReceived |      operations |   10,000,000.00 |   10,000,000.00 |   10,000,000.00 |            0.00 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |           22.59 |           20.20 |           17.74 |            1.28 |
TotalCollections [Gen1] |     collections |            9.86 |            7.74 |            6.48 |            0.82 |
TotalCollections [Gen2] |     collections |            2.55 |            2.03 |            1.36 |            0.37 |
[Counter] MessageReceived |      operations |    4,107,453.28 |    3,778,396.57 |    3,412,137.21 |      226,468.61 |

## After
# Akka.Tests.Performance.Actor.ReceiveActorThroughputSpec+Benchmark
__Measures the throughput of an actor implementation__
_6/12/2016 11:11:28 PM_
### System Info
```ini
NBench=NBench, Version=0.3.0.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=2
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=2
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=13, MaximumRunTime=00:00:01
Concurrent=True
Tracing=True
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |           46.00 |           45.15 |           44.00 |            0.69 |
TotalCollections [Gen1] |     collections |           19.00 |           18.38 |           17.00 |            0.65 |
TotalCollections [Gen2] |     collections |            6.00 |            5.38 |            4.00 |            0.65 |
[Counter] MessageReceived |      operations |   10,000,000.00 |   10,000,000.00 |   10,000,000.00 |            0.00 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |           18.83 |           17.11 |           15.61 |            0.82 |
TotalCollections [Gen1] |     collections |            7.78 |            6.97 |            6.38 |            0.41 |
TotalCollections [Gen2] |     collections |            2.46 |            2.04 |            1.52 |            0.28 |
[Counter] MessageReceived |      operations |    4,092,658.78 |    3,788,223.95 |    3,541,850.70 |      148,096.66 |

The numbers are virtually identical. No real changes there. Both of these tests use the `ThreadPoolDispatcher`.